Extract the tag browser into a per-kit function
The browser was ~365 lines inline in the sidebar closure, addressing the
active kit directly. It now takes the kit as an argument, so a split view can
draw a different kit's browser in each pane rather than one browser that
retargets as focus moves.

Every widget id beneath it is salted with the kit id via push_id. Without
that, two browsers in the same frame would share state for anything egui keys
by label — folder collapse state is keyed on the folder name alone, so
expanding 'objects' in one kit would expand it in the other. Salting at the
top of the function rather than threading a parameter through draw_tree means
a caller cannot forget it.
